refactor change of components, update imports, add preview column
