+++ src/components/InfoIcon/InfoIconWrapper.tsx
@@ -11,8 +11,8 @@
  * SPDX-License-Identifier: EPL-2.0 OR W3C-20150513
  ********************************************************************************/
 import React from "react";
+import Icon from "./Icon";
 import { Info } from "react-feather";
-import { Tooltip } from "react-tooltip";
 
 /**
  * Display an info icon that shows a information text when hovered
@@ -34,34 +34,14 @@ const InfoIconWrapper: React.FC<IInfoIconWrapperProps> = (props) => {
     <div className="flex justify-center">
       <div className="pr-0.5">{props.children}</div>
       <button onClick={() => window.open(props.tooltip.href, "_blank")}>
-        <InfoIcon
+        <Icon
           html={props.tooltip.html || "No tooltip content available"}
           id={props.id}
+          IconComponent={Info}
         />
       </button>
       <div className="p-1"></div>
     </div>
-  );
-};
-
-interface IInfoIconProps {
-  html: string;
-  id: string;
-}
-/**
- * Display an info icon that shows a information text when hovered.
- *
- * The parameter html is a String containing HTML for the tooltips layout.
- * @param {String} html
- */
-export const InfoIcon: React.FC<IInfoIconProps> = (props) => {
-  return (
-    <>
-      <a data-tooltip-id={props.id} data-tooltip-html={props.html}>
-        <Info color="grey" size="16" />
-      </a>
-      <Tooltip id={props.id} place="top" className="z-10" />
-    </>
   );
 };
