+++ src/components/TDViewer/components/EditProperties.tsx
@@ -14,13 +14,12 @@ import React, { useContext, useState } from "react";
 import ediTDorContext from "../../../context/ediTDorContext";
 import ButtonSwap from "../base/ButtonSwap";
 import IncrementButton from "../base/IncrementButton";
-import { IThingDescription } from "types/td";
 import InfoIconWrapper from "../../InfoIcon/InfoIconWrapper";
 import {
   getAddressOffsetTooltipContent,
   getEndiannessTooltipContent,
   getUniIdTooltipContent,
-} from "../../InfoIcon/InfoTooltips";
+} from "../../InfoIcon/TooltipMapper";
 
 interface IEndianness {
   wordSwap: boolean;
